+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Allow URLs with relative protocol (#32)
+
 ## [2.2.0] - 2024-06-11
 
 ### Added